User login/registration JWT implementation. User Controller Endpoints Authentication Updates

+++ src/main/java/com/real/time/chatapp/Controllers/UserController.java
@@ -1,4 +1,4 @@
-package com.real.time.chatapp.Rest;
+package com.real.time.chatapp.Controllers;
 
 import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
 import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
@@ -10,7 +10,12 @@ import java.util.stream.Collectors;
 import org.springframework.hateoas.CollectionModel;
 import org.springframework.hateoas.EntityModel;
 import org.springframework.hateoas.IanaLinkRelations;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -20,27 +25,31 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.real.time.chatapp.Assemblers.ConversationModelAssembler;
+import com.real.time.chatapp.Assemblers.UserModelAssembler;
+import com.real.time.chatapp.Auth.AuthenticationRequest;
+import com.real.time.chatapp.ControllerServices.AuthenticationService;
 import com.real.time.chatapp.Entities.Message;
+import com.real.time.chatapp.Entities.Role;
 import com.real.time.chatapp.Entities.User;
 import com.real.time.chatapp.Exception.UserNotFoundException;
+import com.real.time.chatapp.Repositories.ConversationRepository;
+import com.real.time.chatapp.Repositories.MessageRepository;
+import com.real.time.chatapp.Repositories.UserRepository;
+
+import lombok.RequiredArgsConstructor;
 
 @RestController
-class UserController {
+@RequiredArgsConstructor
+public class UserController {
 
 	private final UserRepository user_repository;
 	private final UserModelAssembler user_assembler;
 	private final ConversationRepository conversation_repository;
 	private final ConversationModelAssembler conversation_assembler;
 	private final MessageRepository message_repository;
-
-	UserController(UserRepository user_repository, UserModelAssembler user_assembler,
-			ConversationModelAssembler conversation_assembler, ConversationRepository conversation_repository, MessageRepository message_repository) {
-		this.user_repository = user_repository;
-		this.user_assembler = user_assembler;
-		this.conversation_repository = conversation_repository;
-		this.conversation_assembler = conversation_assembler;
-		this.message_repository = message_repository;
-	}
+	private final AuthenticationService service;
+	private final AuthenticationManager authenticationManager;
 
 	/**
 	 * Fetching all users
@@ -48,7 +57,7 @@ class UserController {
 	 * @return
 	 */
 	@GetMapping("/users")
-	CollectionModel<EntityModel<User>> all() {
+	public CollectionModel<EntityModel<User>> all() {
 		List<EntityModel<User>> users = user_repository.findAll().stream().map(user_assembler::toModel)
 				.collect(Collectors.toList());
 
@@ -62,14 +71,15 @@ class UserController {
 	 * @return
 	 */
 	@GetMapping("/users/{id}")
-	EntityModel<User> one(@PathVariable Long id) {
+	public EntityModel<User> one(@PathVariable Long id) {
 
 		User user = user_repository.findById(id).orElseThrow(() -> new UserNotFoundException(id));
 		return user_assembler.toModel(user);
 	}
-	
+
 	/**
 	 * Search for users based on name
+	 * 
 	 * @param name
 	 * @return
 	 */
@@ -78,35 +88,47 @@ class UserController {
 		String[] firstAndLast = name.split(" ");
 		List<EntityModel<User>> entityModels = user_repository.searchUsersByName(firstAndLast[0], firstAndLast[1])
 				.stream().map(user_assembler::toModel).collect(Collectors.toList());
-		
+
 		return CollectionModel.of(entityModels,
 				linkTo(methodOn(UserController.class).searchUsersByName(name)).withSelfRel());
 	}
-	
+
 	/**
 	 * Search for users by username
+	 * 
 	 * @param userName
 	 * @return
 	 */
 	@GetMapping("/search/users/userName")
-	CollectionModel<EntityModel<User>> searchUsersByUserName(@RequestParam String userName){
-		List<EntityModel<User>> entityModels = user_repository.searchUsersByUserName(userName).stream().map(user_assembler::toModel).collect(Collectors.toList());
-		
+	CollectionModel<EntityModel<User>> searchUsersByUserName(@RequestParam String userName) {
+		List<EntityModel<User>> entityModels = user_repository.searchUsersByUserName(userName).stream()
+				.map(user_assembler::toModel).collect(Collectors.toList());
+
 		return CollectionModel.of(entityModels,
 				linkTo(methodOn(UserController.class).searchUsersByUserName(userName)).withSelfRel());
 	}
 	
-	
 	/**
-	 * Creating a new User
+	 * Register a User 
 	 * 
-	 * @param newUser
+	 * @param user
 	 * @return
 	 */
-	@PostMapping("/users")
-	ResponseEntity<?> newUser(@RequestBody User newUser) {
-		EntityModel<User> entityModel = user_assembler.toModel(user_repository.save(newUser));
-		return ResponseEntity.created(entityModel.getRequiredLink(IanaLinkRelations.SELF).toUri()).body(entityModel);
+	@PostMapping("/register")
+	ResponseEntity<?> register(@RequestBody User user) {
+		return service.register(user);
+	}
+
+	
+	/**
+	 * Authenticate a User 
+	 * 
+	 * @param request
+	 * @return
+	 */
+	@PostMapping("/authenticate")
+	ResponseEntity<?> authenticate(@RequestBody AuthenticationRequest request){
+		return service.authenticate(request);
 	}
 
 	/**
@@ -118,17 +140,19 @@ class UserController {
 	 */
 	@PutMapping("users/{id}")
 	ResponseEntity<?> updateUser(@RequestBody User newUser, @PathVariable Long id) {
-		User updatedUser = user_repository.findById(id).map(user -> {
-			user.setFirstName(newUser.getFirstName());
-			user.setLastName(newUser.getLastName());
-			user.setPassword(newUser.getPassword());
-			user.setRecievedMessages(newUser.getRecievedMessages());
-			user.setSentMessages(newUser.getSentMessages());
-			user.setUserName(newUser.getUsername());
-			return user_repository.save(user);
-		}).orElseGet(() -> {
-			return user_repository.save(newUser);
-		});
+		User user = user_repository.findById(id).orElseThrow(() -> new UserNotFoundException(id));
+		if(!service.validateUser(user)) {
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("You are not authorized to perform this action.");
+		}
+		user.setFirstName(newUser.getFirstName());
+		user.setLastName(newUser.getLastName());
+		user.setPassword(newUser.getPassword());
+		user.setRecievedMessages(newUser.getRecievedMessages());
+		user.setRole(Role.USER);
+		user.setSentMessages(newUser.getSentMessages());
+		user.setUserName(newUser.getUsername());
+		User updatedUser = user_repository.save(user);
+		
 
 		EntityModel<User> entityModel = user_assembler.toModel(updatedUser);
 
@@ -143,15 +167,20 @@ class UserController {
 	 */
 	@DeleteMapping("/users/{id}")
 	ResponseEntity<?> deleteUser(@PathVariable Long id) {
-		//TODO: Consider another way to do this logic ??
+		// TODO: Consider another way to do this logic ??
 		/**
-		 *  Delete messages sent by User and remove all recipients of each message sent by User
+		 * Delete messages sent by User and remove all recipients of each message sent
+		 * by User
 		 */
-		
+
 		User user = user_repository.findById(id).orElseThrow(() -> new UserNotFoundException(id));
-		for(Message msg: user.getSentMessages()) {
+        //Ensure this user is either an admin or a user deleting their account
+		if(!service.validateUser(user)) {
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("You are not authorized to perform this action.");
+		}
+		for (Message msg : user.getSentMessages()) {
 			Set<User> recipients = msg.getRecipients();
-			for(User current: recipients) {
+			for (User current : recipients) {
 				current.getRecievedMessages().remove(msg);
 			}
 			message_repository.deleteById(msg.getMessage_id());
